+++ CHANGES.md
@@ -1,3 +1,20 @@
+v1.5.0
+
+- revamped hash-collections, thanks again @alex-ozdemir!
+
+    Details in [PR 8](https://github.com/AdrienChampion/hashconsing/pull/8).
+
+    - pre-`v1.5.0` default was the identity hasher, now called `id-hash`;
+    - default hasher is now `p-hash` (faster and deterministic);
+    - Rust's std hash (`sip-hash`) also available (non-deterministic);
+    - `a-hash` available by activating feature `"with_ahash"` (fastest, non-deterministic).
+
+# Breaking Changes (Kinda)
+
+Previous versions were re-exporting `lazy_static`'s main macro, which was technically not needed. It
+has been removed for this version.
+
+
 v1.3.0
 
 - publicized the weak reference API, thanks @alex-ozdemir!
@@ -11,7 +28,7 @@ v1.2.0
 
 # Breaking Changes (Kinda)
 
-The previous version was exposing too much from `lazy_static` which was a mistake. Only the
+Previous versions were exposing too much from `lazy_static` which was a mistake. Only the
 `lazy_static` macro is re-exported now, other items such as the `__Deref` trait are not re-exported
 anymore.